Changed validator

+++ src/main/java/com/catalyst/collector/services/impl/CollectionsServiceImpl.java
@@ -196,6 +196,8 @@ public class CollectionsServiceImpl implements CollectionsService {
 
     @Override
     public void deleteCondition(Integer id) {
+        if (id < 0)
+            return;
         collectionsDao.deleteCondition(id);
     }
